+++ compiler/tests-compiler/js_parser_printer.ml
@@ -1389,8 +1389,8 @@ let%expect_test "in operator in conditional then-branch needs no parentheses" =
     |}]
 
 (* [async] and [using] are contextual keywords: they only introduce a
-   restricted production when followed by specific tokens ([function], [*],
-   arrow parameters, a class element name, a binding).  Used as plain
+   restricted production when followed by specific tokens ([function], an
+   arrow parameter name, a class element name, a binding).  Used as plain
    identifiers, a line break after them must not trigger ASI. *)
 let%expect_test "async/using identifiers followed by a line break" =
   let test s = print ~debuginfo:false ~report:true ~compact:false s in
@@ -1399,33 +1399,83 @@ var async = 1;
 var x = async
 + 1;
 |};
-  [%expect {| var async = 1; var x = async; + 1; |}];
+  [%expect {| var async = 1; var x = async + 1; |}];
   test {|
 var async
 = 1;
 |};
-  [%expect {| cannot parse js (from l:3, c:0)@. |}];
+  [%expect {| var async = 1; |}];
   test {|
 var async = { f() { return 1 } };
 var x = async
 .f();
 |};
-  [%expect {| cannot parse js (from l:4, c:0)@. |}];
+  [%expect {| var async = {f(){return 1;}}; var x = async.f(); |}];
   test {|
 var async = (a) => a;
 var x = async
 (2);
 |};
-  [%expect {| var async = a=>a; var x = async; 2; |}];
+  [%expect {| var async = a=>a; var x = async(2); |}];
   test {|
 var using = 1;
 var x = using
 + 1;
 |};
-  [%expect {| var using = 1; var x = using; + 1; |}];
+  [%expect {| var using = 1; var x = using + 1; |}];
   test {|
 var using = (a) => a;
 var x = using
 (2);
 |};
-  [%expect {| var using = a=>a; var x = using; 2; |}]
+  [%expect {| var using = a=>a; var x = using(2); |}];
+  (* [*] is only restricted in method definitions *)
+  test {|
+var async = 1;
+var x = async
+* 2;
+|};
+  [%expect {| var async = 1; var x = async * 2; |}];
+  test {|
+var async = 1;
+async
+: 1;
+|};
+  [%expect {| var async = 1; async: 1; |}];
+  (* Restricted continuations: a semicolon must still be inserted. *)
+  test {|
+var x = async
+function f() {}
+|};
+  [%expect {| var x = async; function f(){} |}];
+  test {|
+var x = async
+y => 1;
+|};
+  [%expect {| var x = async; y=>1; |}];
+  test {|
+var using = 1;
+using
+x = 2;
+|};
+  [%expect {| var using = 1; using; x = 2; |}];
+  (* In a class body, [async] on its own line is a field name, unless
+     directly followed by [(], [=] or [;]. *)
+  test
+    {|
+class A {
+  async
+  foo() {}
+  async
+  * gen() {}
+  async
+  [x]() {}
+  async
+  (x) {}
+  async
+  = 1;
+  async
+}
+|};
+  [%expect
+    {| class A{async;foo(){}async;* gen(){}async;[x](){}async(x){}async = 1;async;} |}]